Update styles and dialogs

+++ packages/menu/src/lib/components/categories-table/categories-table.component.ts
@@ -1,18 +1,21 @@
 import { CdkDrag, CdkDragDrop, CdkDropList, moveItemInArray } from '@angular/cdk/drag-drop';
 import { Component, EventEmitter, Input, Output, inject } from '@angular/core';
 import { Router } from '@angular/router';
-import { TranslateModule } from '@ngx-translate/core';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { CustomMessageService } from 'theme-shared';
 import { tap } from 'rxjs';
 import { CategoryModel } from '../../models/category.model';
 import { MenuService } from '../../services/menu.service';
+import { ConfirmDialogModule } from 'primeng/confirmdialog';
+import { ConfirmationService } from 'primeng/api';
 
 @Component({
   selector: 'app-menu-categories-table',
   standalone: true,
   templateUrl: './categories-table.component.html',
   styleUrl: './categories-table.component.scss',
-  imports: [CdkDropList, CdkDrag, TranslateModule],
+  imports: [CdkDropList, CdkDrag, TranslateModule, ConfirmDialogModule],
+  providers: [ConfirmationService]
 })
 export class CategoriesTableComponent {
   @Input() categories: CategoryModel[] = [];
@@ -22,6 +25,8 @@ export class CategoriesTableComponent {
   router = inject(Router);
   menuService = inject(MenuService);
   customMessageService = inject(CustomMessageService);
+  confirmationService = inject(ConfirmationService);
+  translateService = inject(TranslateService);
 
   drop(event: CdkDragDrop<string[]>) {
     moveItemInArray(this.categories, event.previousIndex, event.currentIndex);
@@ -35,14 +40,22 @@ export class CategoriesTableComponent {
   }
 
   deleteCategory(deletedCategory: CategoryModel) {
-    this.menuService
-      .deleteCategory(deletedCategory._id)
-      .pipe(
-        tap((res) => {
-          this.updateCategoriesList.emit();
-          this.customMessageService.success(res);
-        }),
-      )
-      .subscribe();
+    this.confirmationService.confirm({
+      message: this.translateService.instant('menu.deleteCategoryMessage', {
+        name: deletedCategory.name,
+      }),
+      header: this.translateService.instant('menu.deleteCategoryHeader'),
+      accept: () => {
+        this.menuService
+          .deleteCategory(deletedCategory._id)
+          .pipe(
+            tap((res) => {
+              this.updateCategoriesList.emit();
+              this.customMessageService.success(res);
+            }),
+          )
+          .subscribe();
+      },
+    });
   }
 }
