+++ packages/menu/src/lib/components/products-table/products-table.component.ts
@@ -2,12 +2,14 @@ import { CurrencyPipe, NgClass } from '@angular/common';
 import { Component, EventEmitter, Input, Output, inject } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { CdkDrag, CdkDragDrop, CdkDropList, moveItemInArray } from '@angular/cdk/drag-drop';
-import { TranslateModule } from '@ngx-translate/core';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { ProductModel } from '../../models/product.model';
 import { MenuService } from '../../services';
 import { tap } from 'rxjs';
 import { EditProductComponent } from '../edit-product-modal/edit-product.component';
 import { CustomMessageService } from 'theme-shared';
+import { ConfirmationService } from 'primeng/api';
+import { ConfirmDialogModule } from 'primeng/confirmdialog'
 
 @Component({
   selector: 'app-products-table',
@@ -22,14 +24,18 @@ import { CustomMessageService } from 'theme-shared';
     TranslateModule,
     NgClass,
     EditProductComponent,
+    ConfirmDialogModule
   ],
+  providers:[ConfirmationService]
 })
 export class ProductsTableComponent {
   @Input() products: ProductModel[];
   @Output() updateProductsPlacement = new EventEmitter<ProductModel[]>();
   @Output() updateList = new EventEmitter<boolean>();
-  menuService = inject(MenuService);
-  customMessageService = inject(CustomMessageService);
+  protected readonly menuService = inject(MenuService);
+  protected readonly customMessageService = inject(CustomMessageService);
+  protected readonly confirmationService = inject(ConfirmationService);
+  protected readonly translateService = inject(TranslateService);
 
   visibleEditProductDialog = false;
   product: ProductModel;
@@ -42,16 +48,25 @@ export class ProductsTableComponent {
   }
 
   deleteProduct(product: ProductModel) {
-    this.menuService
-      .deleteProduct(product._id)
-      .pipe(
-        tap((res) => {
-          this.customMessageService.success(res);
-          this.updateList.emit(true);
-        }),
-      )
-      .subscribe();
+    this.confirmationService.confirm({
+      message: this.translateService.instant('menu.deleteProductMessage', {
+        name: product.name,
+      }),
+      header: this.translateService.instant('menu.deleteProductHeader'),
+      accept: () => {
+        this.menuService
+        .deleteProduct(product._id)
+        .pipe(
+          tap((res) => {
+            this.customMessageService.success(res);
+            this.updateList.emit(true);
+          }),
+        )
+        .subscribe();
+      },
+    });
   }
+
 
   openEditProductModal(product: ProductModel) {
     this.visibleEditProductDialog = true;
